Extract code dealing with native-keymap in its own file
